requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
fire==0.3.1
IPy==1.1
python-nmap==0.6.1
urllib3==1.25.11
pandas==1.2.0
sqlalchemy==1.3.20
beautifulsoup4==4.9.3
bs4==0.0.1
loguru==0.5.3
requests==2.24.0
tqdm==4.51.0
